The enums have been split into external files.
