+++ src/System.Windows.Forms/src/System/Windows/Forms/ComboBox.ChildWindowType.cs
@@ -1,0 +1,19 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Windows.Forms
+{
+    public partial class ComboBox
+    {
+        /// <summary>
+        ///  Represents the ComboBox child native window type.
+        /// </summary>
+        private enum ChildWindowType
+        {
+            ListBox,
+            Edit,
+            DropDownList
+        }
+    }
+}
