+++ src/System.Windows.Forms/src/System/Windows/Forms/Layout/CommonProperties.DockAnchorMode.cs
@@ -1,0 +1,15 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Windows.Forms.Layout
+{
+    internal partial class CommonProperties
+    {
+        private enum DockAnchorMode
+        {
+            Anchor = 0,
+            Dock = 1
+        }
+    }
+}
