+++ src/System.Windows.Forms/src/System/Windows/Forms/ProfessionalColorTable.KnownColors.cs
@@ -1,0 +1,227 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Windows.Forms
+{
+    public partial class ProfessionalColorTable
+    {
+        private enum KnownColors
+        {
+            msocbvcrCBBdrOuterDocked,
+            msocbvcrCBBdrOuterFloating,
+            msocbvcrCBBkgd,
+            msocbvcrCBCtlBdrMouseDown,
+            msocbvcrCBCtlBdrMouseOver,
+            msocbvcrCBCtlBdrSelected,
+            msocbvcrCBCtlBdrSelectedMouseOver,
+            msocbvcrCBCtlBkgd,
+            msocbvcrCBCtlBkgdLight,
+            msocbvcrCBCtlBkgdMouseDown,
+            msocbvcrCBCtlBkgdMouseOver,
+            msocbvcrCBCtlBkgdSelected,
+            msocbvcrCBCtlBkgdSelectedMouseOver,
+            msocbvcrCBCtlText,
+            msocbvcrCBCtlTextDisabled,
+            msocbvcrCBCtlTextLight,
+            msocbvcrCBCtlTextMouseDown,
+            msocbvcrCBCtlTextMouseOver,
+            msocbvcrCBDockSeparatorLine,
+            msocbvcrCBDragHandle,
+            msocbvcrCBDragHandleShadow,
+            msocbvcrCBDropDownArrow,
+            msocbvcrCBGradMainMenuHorzBegin,
+            msocbvcrCBGradMainMenuHorzEnd,
+            msocbvcrCBGradMenuIconBkgdDroppedBegin,
+            msocbvcrCBGradMenuIconBkgdDroppedEnd,
+            msocbvcrCBGradMenuIconBkgdDroppedMiddle,
+            msocbvcrCBGradMenuTitleBkgdBegin,
+            msocbvcrCBGradMenuTitleBkgdEnd,
+            msocbvcrCBGradMouseDownBegin,
+            msocbvcrCBGradMouseDownEnd,
+            msocbvcrCBGradMouseDownMiddle,
+            msocbvcrCBGradMouseOverBegin,
+            msocbvcrCBGradMouseOverEnd,
+            msocbvcrCBGradMouseOverMiddle,
+            msocbvcrCBGradOptionsBegin,
+            msocbvcrCBGradOptionsEnd,
+            msocbvcrCBGradOptionsMiddle,
+            msocbvcrCBGradOptionsMouseOverBegin,
+            msocbvcrCBGradOptionsMouseOverEnd,
+            msocbvcrCBGradOptionsMouseOverMiddle,
+            msocbvcrCBGradOptionsSelectedBegin,
+            msocbvcrCBGradOptionsSelectedEnd,
+            msocbvcrCBGradOptionsSelectedMiddle,
+            msocbvcrCBGradSelectedBegin,
+            msocbvcrCBGradSelectedEnd,
+            msocbvcrCBGradSelectedMiddle,
+            msocbvcrCBGradVertBegin,
+            msocbvcrCBGradVertEnd,
+            msocbvcrCBGradVertMiddle,
+            msocbvcrCBIconDisabledDark,
+            msocbvcrCBIconDisabledLight,
+            msocbvcrCBLabelBkgnd,
+            msocbvcrCBLowColorIconDisabled,
+            msocbvcrCBMainMenuBkgd,
+            msocbvcrCBMenuBdrOuter,
+            msocbvcrCBMenuBkgd,
+            msocbvcrCBMenuCtlText,
+            msocbvcrCBMenuCtlTextDisabled,
+            msocbvcrCBMenuIconBkgd,
+            msocbvcrCBMenuIconBkgdDropped,
+            msocbvcrCBMenuShadow,
+            msocbvcrCBMenuSplitArrow,
+            msocbvcrCBOptionsButtonShadow,
+            msocbvcrCBShadow,
+            msocbvcrCBSplitterLine,
+            msocbvcrCBSplitterLineLight,
+            msocbvcrCBTearOffHandle,
+            msocbvcrCBTearOffHandleMouseOver,
+            msocbvcrCBTitleBkgd,
+            msocbvcrCBTitleText,
+            msocbvcrDisabledFocuslessHighlightedText,
+            msocbvcrDisabledHighlightedText,
+            msocbvcrDlgGroupBoxText,
+            msocbvcrDocTabBdr,
+            msocbvcrDocTabBdrDark,
+            msocbvcrDocTabBdrDarkMouseDown,
+            msocbvcrDocTabBdrDarkMouseOver,
+            msocbvcrDocTabBdrLight,
+            msocbvcrDocTabBdrLightMouseDown,
+            msocbvcrDocTabBdrLightMouseOver,
+            msocbvcrDocTabBdrMouseDown,
+            msocbvcrDocTabBdrMouseOver,
+            msocbvcrDocTabBdrSelected,
+            msocbvcrDocTabBkgd,
+            msocbvcrDocTabBkgdMouseDown,
+            msocbvcrDocTabBkgdMouseOver,
+            msocbvcrDocTabBkgdSelected,
+            msocbvcrDocTabText,
+            msocbvcrDocTabTextMouseDown,
+            msocbvcrDocTabTextMouseOver,
+            msocbvcrDocTabTextSelected,
+            msocbvcrDWActiveTabBkgd,
+            msocbvcrDWActiveTabText,
+            msocbvcrDWActiveTabTextDisabled,
+            msocbvcrDWInactiveTabBkgd,
+            msocbvcrDWInactiveTabText,
+            msocbvcrDWTabBkgdMouseDown,
+            msocbvcrDWTabBkgdMouseOver,
+            msocbvcrDWTabTextMouseDown,
+            msocbvcrDWTabTextMouseOver,
+            msocbvcrFocuslessHighlightedBkgd,
+            msocbvcrFocuslessHighlightedText,
+            msocbvcrGDHeaderBdr,
+            msocbvcrGDHeaderBkgd,
+            msocbvcrGDHeaderCellBdr,
+            msocbvcrGDHeaderCellBkgd,
+            msocbvcrGDHeaderCellBkgdSelected,
+            msocbvcrGDHeaderSeeThroughSelection,
+            msocbvcrGSPDarkBkgd,
+            msocbvcrGSPGroupContentDarkBkgd,
+            msocbvcrGSPGroupContentLightBkgd,
+            msocbvcrGSPGroupContentText,
+            msocbvcrGSPGroupContentTextDisabled,
+            msocbvcrGSPGroupHeaderDarkBkgd,
+            msocbvcrGSPGroupHeaderLightBkgd,
+            msocbvcrGSPGroupHeaderText,
+            msocbvcrGSPGroupline,
+            msocbvcrGSPHyperlink,
+            msocbvcrGSPLightBkgd,
+            msocbvcrHyperlink,
+            msocbvcrHyperlinkFollowed,
+            msocbvcrJotNavUIBdr,
+            msocbvcrJotNavUIGradBegin,
+            msocbvcrJotNavUIGradEnd,
+            msocbvcrJotNavUIGradMiddle,
+            msocbvcrJotNavUIText,
+            msocbvcrListHeaderArrow,
+            msocbvcrNetLookBkgnd,
+            msocbvcrOABBkgd,
+            msocbvcrOBBkgdBdr,
+            msocbvcrOBBkgdBdrContrast,
+            msocbvcrOGMDIParentWorkspaceBkgd,
+            msocbvcrOGRulerActiveBkgd,
+            msocbvcrOGRulerBdr,
+            msocbvcrOGRulerBkgd,
+            msocbvcrOGRulerInactiveBkgd,
+            msocbvcrOGRulerTabBoxBdr,
+            msocbvcrOGRulerTabBoxBdrHighlight,
+            msocbvcrOGRulerTabStopTicks,
+            msocbvcrOGRulerText,
+            msocbvcrOGTaskPaneGroupBoxHeaderBkgd,
+            msocbvcrOGWorkspaceBkgd,
+            msocbvcrOLKFlagNone,
+            msocbvcrOLKFolderbarDark,
+            msocbvcrOLKFolderbarLight,
+            msocbvcrOLKFolderbarText,
+            msocbvcrOLKGridlines,
+            msocbvcrOLKGroupLine,
+            msocbvcrOLKGroupNested,
+            msocbvcrOLKGroupShaded,
+            msocbvcrOLKGroupText,
+            msocbvcrOLKIconBar,
+            msocbvcrOLKInfoBarBkgd,
+            msocbvcrOLKInfoBarText,
+            msocbvcrOLKPreviewPaneLabelText,
+            msocbvcrOLKTodayIndicatorDark,
+            msocbvcrOLKTodayIndicatorLight,
+            msocbvcrOLKWBActionDividerLine,
+            msocbvcrOLKWBButtonDark,
+            msocbvcrOLKWBButtonLight,
+            msocbvcrOLKWBDarkOutline,
+            msocbvcrOLKWBFoldersBackground,
+            msocbvcrOLKWBHoverButtonDark,
+            msocbvcrOLKWBHoverButtonLight,
+            msocbvcrOLKWBLabelText,
+            msocbvcrOLKWBPressedButtonDark,
+            msocbvcrOLKWBPressedButtonLight,
+            msocbvcrOLKWBSelectedButtonDark,
+            msocbvcrOLKWBSelectedButtonLight,
+            msocbvcrOLKWBSplitterDark,
+            msocbvcrOLKWBSplitterLight,
+            msocbvcrPlacesBarBkgd,
+            msocbvcrPPOutlineThumbnailsPaneTabAreaBkgd,
+            msocbvcrPPOutlineThumbnailsPaneTabBdr,
+            msocbvcrPPOutlineThumbnailsPaneTabInactiveBkgd,
+            msocbvcrPPOutlineThumbnailsPaneTabText,
+            msocbvcrPPSlideBdrActiveSelected,
+            msocbvcrPPSlideBdrActiveSelectedMouseOver,
+            msocbvcrPPSlideBdrInactiveSelected,
+            msocbvcrPPSlideBdrMouseOver,
+            msocbvcrPubPrintDocScratchPageBkgd,
+            msocbvcrPubWebDocScratchPageBkgd,
+            msocbvcrSBBdr,
+            msocbvcrScrollbarBkgd,
+            msocbvcrToastGradBegin,
+            msocbvcrToastGradEnd,
+            msocbvcrWPBdrInnerDocked,
+            msocbvcrWPBdrOuterDocked,
+            msocbvcrWPBdrOuterFloating,
+            msocbvcrWPBkgd,
+            msocbvcrWPCtlBdr,
+            msocbvcrWPCtlBdrDefault,
+            msocbvcrWPCtlBdrDisabled,
+            msocbvcrWPCtlBkgd,
+            msocbvcrWPCtlBkgdDisabled,
+            msocbvcrWPCtlText,
+            msocbvcrWPCtlTextDisabled,
+            msocbvcrWPCtlTextMouseDown,
+            msocbvcrWPGroupline,
+            msocbvcrWPInfoTipBkgd,
+            msocbvcrWPInfoTipText,
+            msocbvcrWPNavBarBkgnd,
+            msocbvcrWPText,
+            msocbvcrWPTextDisabled,
+            msocbvcrWPTitleBkgdActive,
+            msocbvcrWPTitleBkgdInactive,
+            msocbvcrWPTitleTextActive,
+            msocbvcrWPTitleTextInactive,
+            msocbvcrXLFormulaBarBkgd,
+            ButtonSelectedHighlight, // not actually from MSO tables
+            ButtonPressedHighlight,// not actually from MSO tables
+            ButtonCheckedHighlight,// not actually from MSO tables
+            lastKnownColor = ButtonCheckedHighlight
+        }
+    }
+}
