+++ src/System.Windows.Forms/src/System/Windows/Forms/ProfessionalColorTable.cs
@@ -7,7 +7,7 @@ using System.Windows.Forms.VisualStyles;
 
 namespace System.Windows.Forms
 {
-    public class ProfessionalColorTable
+    public partial class ProfessionalColorTable
     {
         private Dictionary<KnownColors, Color>? _professionalRGB;
         private bool _usingSystemColors;
@@ -1676,224 +1676,6 @@ namespace System.Windows.Forms
             rgbTable[ProfessionalColorTable.KnownColors.msocbvcrWPTitleTextActive] = Color.FromArgb(0, 0, 0);
             rgbTable[ProfessionalColorTable.KnownColors.msocbvcrWPTitleTextInactive] = Color.FromArgb(0, 0, 0);
             rgbTable[ProfessionalColorTable.KnownColors.msocbvcrXLFormulaBarBkgd] = Color.FromArgb(158, 190, 245);
-        }
-
-        private enum KnownColors
-        {
-            msocbvcrCBBdrOuterDocked,
-            msocbvcrCBBdrOuterFloating,
-            msocbvcrCBBkgd,
-            msocbvcrCBCtlBdrMouseDown,
-            msocbvcrCBCtlBdrMouseOver,
-            msocbvcrCBCtlBdrSelected,
-            msocbvcrCBCtlBdrSelectedMouseOver,
-            msocbvcrCBCtlBkgd,
-            msocbvcrCBCtlBkgdLight,
-            msocbvcrCBCtlBkgdMouseDown,
-            msocbvcrCBCtlBkgdMouseOver,
-            msocbvcrCBCtlBkgdSelected,
-            msocbvcrCBCtlBkgdSelectedMouseOver,
-            msocbvcrCBCtlText,
-            msocbvcrCBCtlTextDisabled,
-            msocbvcrCBCtlTextLight,
-            msocbvcrCBCtlTextMouseDown,
-            msocbvcrCBCtlTextMouseOver,
-            msocbvcrCBDockSeparatorLine,
-            msocbvcrCBDragHandle,
-            msocbvcrCBDragHandleShadow,
-            msocbvcrCBDropDownArrow,
-            msocbvcrCBGradMainMenuHorzBegin,
-            msocbvcrCBGradMainMenuHorzEnd,
-            msocbvcrCBGradMenuIconBkgdDroppedBegin,
-            msocbvcrCBGradMenuIconBkgdDroppedEnd,
-            msocbvcrCBGradMenuIconBkgdDroppedMiddle,
-            msocbvcrCBGradMenuTitleBkgdBegin,
-            msocbvcrCBGradMenuTitleBkgdEnd,
-            msocbvcrCBGradMouseDownBegin,
-            msocbvcrCBGradMouseDownEnd,
-            msocbvcrCBGradMouseDownMiddle,
-            msocbvcrCBGradMouseOverBegin,
-            msocbvcrCBGradMouseOverEnd,
-            msocbvcrCBGradMouseOverMiddle,
-            msocbvcrCBGradOptionsBegin,
-            msocbvcrCBGradOptionsEnd,
-            msocbvcrCBGradOptionsMiddle,
-            msocbvcrCBGradOptionsMouseOverBegin,
-            msocbvcrCBGradOptionsMouseOverEnd,
-            msocbvcrCBGradOptionsMouseOverMiddle,
-            msocbvcrCBGradOptionsSelectedBegin,
-            msocbvcrCBGradOptionsSelectedEnd,
-            msocbvcrCBGradOptionsSelectedMiddle,
-            msocbvcrCBGradSelectedBegin,
-            msocbvcrCBGradSelectedEnd,
-            msocbvcrCBGradSelectedMiddle,
-            msocbvcrCBGradVertBegin,
-            msocbvcrCBGradVertEnd,
-            msocbvcrCBGradVertMiddle,
-            msocbvcrCBIconDisabledDark,
-            msocbvcrCBIconDisabledLight,
-            msocbvcrCBLabelBkgnd,
-            msocbvcrCBLowColorIconDisabled,
-            msocbvcrCBMainMenuBkgd,
-            msocbvcrCBMenuBdrOuter,
-            msocbvcrCBMenuBkgd,
-            msocbvcrCBMenuCtlText,
-            msocbvcrCBMenuCtlTextDisabled,
-            msocbvcrCBMenuIconBkgd,
-            msocbvcrCBMenuIconBkgdDropped,
-            msocbvcrCBMenuShadow,
-            msocbvcrCBMenuSplitArrow,
-            msocbvcrCBOptionsButtonShadow,
-            msocbvcrCBShadow,
-            msocbvcrCBSplitterLine,
-            msocbvcrCBSplitterLineLight,
-            msocbvcrCBTearOffHandle,
-            msocbvcrCBTearOffHandleMouseOver,
-            msocbvcrCBTitleBkgd,
-            msocbvcrCBTitleText,
-            msocbvcrDisabledFocuslessHighlightedText,
-            msocbvcrDisabledHighlightedText,
-            msocbvcrDlgGroupBoxText,
-            msocbvcrDocTabBdr,
-            msocbvcrDocTabBdrDark,
-            msocbvcrDocTabBdrDarkMouseDown,
-            msocbvcrDocTabBdrDarkMouseOver,
-            msocbvcrDocTabBdrLight,
-            msocbvcrDocTabBdrLightMouseDown,
-            msocbvcrDocTabBdrLightMouseOver,
-            msocbvcrDocTabBdrMouseDown,
-            msocbvcrDocTabBdrMouseOver,
-            msocbvcrDocTabBdrSelected,
-            msocbvcrDocTabBkgd,
-            msocbvcrDocTabBkgdMouseDown,
-            msocbvcrDocTabBkgdMouseOver,
-            msocbvcrDocTabBkgdSelected,
-            msocbvcrDocTabText,
-            msocbvcrDocTabTextMouseDown,
-            msocbvcrDocTabTextMouseOver,
-            msocbvcrDocTabTextSelected,
-            msocbvcrDWActiveTabBkgd,
-            msocbvcrDWActiveTabText,
-            msocbvcrDWActiveTabTextDisabled,
-            msocbvcrDWInactiveTabBkgd,
-            msocbvcrDWInactiveTabText,
-            msocbvcrDWTabBkgdMouseDown,
-            msocbvcrDWTabBkgdMouseOver,
-            msocbvcrDWTabTextMouseDown,
-            msocbvcrDWTabTextMouseOver,
-            msocbvcrFocuslessHighlightedBkgd,
-            msocbvcrFocuslessHighlightedText,
-            msocbvcrGDHeaderBdr,
-            msocbvcrGDHeaderBkgd,
-            msocbvcrGDHeaderCellBdr,
-            msocbvcrGDHeaderCellBkgd,
-            msocbvcrGDHeaderCellBkgdSelected,
-            msocbvcrGDHeaderSeeThroughSelection,
-            msocbvcrGSPDarkBkgd,
-            msocbvcrGSPGroupContentDarkBkgd,
-            msocbvcrGSPGroupContentLightBkgd,
-            msocbvcrGSPGroupContentText,
-            msocbvcrGSPGroupContentTextDisabled,
-            msocbvcrGSPGroupHeaderDarkBkgd,
-            msocbvcrGSPGroupHeaderLightBkgd,
-            msocbvcrGSPGroupHeaderText,
-            msocbvcrGSPGroupline,
-            msocbvcrGSPHyperlink,
-            msocbvcrGSPLightBkgd,
-            msocbvcrHyperlink,
-            msocbvcrHyperlinkFollowed,
-            msocbvcrJotNavUIBdr,
-            msocbvcrJotNavUIGradBegin,
-            msocbvcrJotNavUIGradEnd,
-            msocbvcrJotNavUIGradMiddle,
-            msocbvcrJotNavUIText,
-            msocbvcrListHeaderArrow,
-            msocbvcrNetLookBkgnd,
-            msocbvcrOABBkgd,
-            msocbvcrOBBkgdBdr,
-            msocbvcrOBBkgdBdrContrast,
-            msocbvcrOGMDIParentWorkspaceBkgd,
-            msocbvcrOGRulerActiveBkgd,
-            msocbvcrOGRulerBdr,
-            msocbvcrOGRulerBkgd,
-            msocbvcrOGRulerInactiveBkgd,
-            msocbvcrOGRulerTabBoxBdr,
-            msocbvcrOGRulerTabBoxBdrHighlight,
-            msocbvcrOGRulerTabStopTicks,
-            msocbvcrOGRulerText,
-            msocbvcrOGTaskPaneGroupBoxHeaderBkgd,
-            msocbvcrOGWorkspaceBkgd,
-            msocbvcrOLKFlagNone,
-            msocbvcrOLKFolderbarDark,
-            msocbvcrOLKFolderbarLight,
-            msocbvcrOLKFolderbarText,
-            msocbvcrOLKGridlines,
-            msocbvcrOLKGroupLine,
-            msocbvcrOLKGroupNested,
-            msocbvcrOLKGroupShaded,
-            msocbvcrOLKGroupText,
-            msocbvcrOLKIconBar,
-            msocbvcrOLKInfoBarBkgd,
-            msocbvcrOLKInfoBarText,
-            msocbvcrOLKPreviewPaneLabelText,
-            msocbvcrOLKTodayIndicatorDark,
-            msocbvcrOLKTodayIndicatorLight,
-            msocbvcrOLKWBActionDividerLine,
-            msocbvcrOLKWBButtonDark,
-            msocbvcrOLKWBButtonLight,
-            msocbvcrOLKWBDarkOutline,
-            msocbvcrOLKWBFoldersBackground,
-            msocbvcrOLKWBHoverButtonDark,
-            msocbvcrOLKWBHoverButtonLight,
-            msocbvcrOLKWBLabelText,
-            msocbvcrOLKWBPressedButtonDark,
-            msocbvcrOLKWBPressedButtonLight,
-            msocbvcrOLKWBSelectedButtonDark,
-            msocbvcrOLKWBSelectedButtonLight,
-            msocbvcrOLKWBSplitterDark,
-            msocbvcrOLKWBSplitterLight,
-            msocbvcrPlacesBarBkgd,
-            msocbvcrPPOutlineThumbnailsPaneTabAreaBkgd,
-            msocbvcrPPOutlineThumbnailsPaneTabBdr,
-            msocbvcrPPOutlineThumbnailsPaneTabInactiveBkgd,
-            msocbvcrPPOutlineThumbnailsPaneTabText,
-            msocbvcrPPSlideBdrActiveSelected,
-            msocbvcrPPSlideBdrActiveSelectedMouseOver,
-            msocbvcrPPSlideBdrInactiveSelected,
-            msocbvcrPPSlideBdrMouseOver,
-            msocbvcrPubPrintDocScratchPageBkgd,
-            msocbvcrPubWebDocScratchPageBkgd,
-            msocbvcrSBBdr,
-            msocbvcrScrollbarBkgd,
-            msocbvcrToastGradBegin,
-            msocbvcrToastGradEnd,
-            msocbvcrWPBdrInnerDocked,
-            msocbvcrWPBdrOuterDocked,
-            msocbvcrWPBdrOuterFloating,
-            msocbvcrWPBkgd,
-            msocbvcrWPCtlBdr,
-            msocbvcrWPCtlBdrDefault,
-            msocbvcrWPCtlBdrDisabled,
-            msocbvcrWPCtlBkgd,
-            msocbvcrWPCtlBkgdDisabled,
-            msocbvcrWPCtlText,
-            msocbvcrWPCtlTextDisabled,
-            msocbvcrWPCtlTextMouseDown,
-            msocbvcrWPGroupline,
-            msocbvcrWPInfoTipBkgd,
-            msocbvcrWPInfoTipText,
-            msocbvcrWPNavBarBkgnd,
-            msocbvcrWPText,
-            msocbvcrWPTextDisabled,
-            msocbvcrWPTitleBkgdActive,
-            msocbvcrWPTitleBkgdInactive,
-            msocbvcrWPTitleTextActive,
-            msocbvcrWPTitleTextInactive,
-            msocbvcrXLFormulaBarBkgd,
-            ButtonSelectedHighlight, // not actually from MSO tables
-            ButtonPressedHighlight,// not actually from MSO tables
-            ButtonCheckedHighlight,// not actually from MSO tables
-            lastKnownColor = ButtonCheckedHighlight
         }
     }
 }
