+++ src/System.Windows.Forms/src/System/Windows/Forms/Triangle.TriangleDirection.cs
@@ -1,0 +1,17 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Windows.Forms
+{
+    internal static partial class Triangle
+    {
+        internal enum TriangleDirection
+        {
+            Up,
+            Down,
+            Left,
+            Right
+        }
+    }
+}
